+++ packages/common/e2e/fixtures/resetConfig.ts
@@ -1,0 +1,272 @@
+export const resetConfig = {
+  orgs: [
+    {
+      orgShort: "prod-pre-p-hub",
+      url: "https://prod-pre-p-hub.maps.arcgis.com",
+      users: [
+        "e2e_pre_pub_admin",
+        "e2e_pre_pub_collaborator",
+        "e2e_pre_pub_publisher",
+      ],
+    },
+    {
+      orgShort: "prod-pre-hub-c",
+      url: "https://prod-pre-hub-c-hub.maps.arcgis.com",
+      users: ["e2e_pre_hub_c_member"],
+    },
+    {
+      orgShort: "prod-bas-p-hub",
+      url: "https://prod-bas-p-hub.maps.arcgis.com",
+      users: [
+        "e2e_bas_priv_admin",
+        "e2e_bas_priv_collaborator",
+        "e2e_bas_priv_publisher",
+      ],
+    },
+    {
+      orgShort: "prod-bas-hub",
+      url: "https://prod-bas-hub.maps.arcgis.com",
+      users: [
+        "e2e_bas_pub_admin",
+        "e2e_bas_pub_collaborator",
+        "e2e_bas_pub_publisher",
+      ],
+    },
+    {
+      orgShort: "qa-pre-p-hub",
+      url: "https://qa-pre-p-hub.mapsqa.arcgis.com",
+      users: [
+        "e2e_pre_priv_admin",
+        "e2e_pre_priv_collaborator",
+        "e2e_pre_priv_publisher",
+      ],
+    },
+    {
+      orgShort: "qa-pre-a-hub",
+      url: "https://qa-pre-a-hub.mapsqa.arcgis.com",
+      users: [
+        "e2e_pre_a_hub_admin",
+        "e2e_pre_a_hub_publisher",
+        "e2e_pre_a_hub_discussions_admin",
+        "e2e_pre_a_hub_discussions_initiative_manager",
+        "e2e_pre_a_hub_discussions_user",
+      ],
+    },
+    {
+      orgShort: "qa-pre-hub-c",
+      url: "https://qa-pre-hub-c.mapsqa.arcgis.com",
+      users: ["e2e_pre_hub_c_member"],
+    },
+    {
+      orgShort: "qa-pre-hub",
+      url: "https://qa-pre-hub.mapsqa.arcgis.com",
+      users: [
+        "e2e_pre_pub_admin",
+        "e2e_pre_pub_collaborator",
+        "e2e_pre_pub_publisher",
+      ],
+    },
+    {
+      orgShort: "qa-bas-p-hub",
+      url: "https://qa-bas-p-hub.mapsqa.arcgis.com",
+      users: [
+        "e2e_bas_priv_admin",
+        "e2e_bas_priv_collaborator",
+        "e2e_bas_priv_publisher",
+      ],
+    },
+    {
+      orgShort: "qa-bas-hub",
+      url: "https://qa-bas-hub.mapsqa.arcgis.com",
+      users: [
+        "e2e_bas_pub_admin",
+        "e2e_bas_pub_collaborator",
+        "e2e_bas_pub_publisher",
+      ],
+    },
+    {
+      orgShort: "dev-pre-p-hub",
+      url: "https://dev-pre-p-hub.mapsdevext.arcgis.com",
+      users: [
+        "e2e_pre_priv_admin",
+        "e2e_pre_priv_collaborator",
+        "e2e_pre_priv_publisher",
+      ],
+    },
+    {
+      orgShort: "dev-pre-hub-c",
+      url: "https://dev-pre-hub-c.mapsdevext.arcgis.com",
+      users: ["e2e_pre_hub_c_member"],
+    },
+    {
+      orgShort: "dev-pre-hub",
+      url: "https://dev-pre-hub.mapsdevext.arcgis.com",
+      users: [
+        "e2e_pre_pub_admin",
+        "e2e_pre_pub_collaborator",
+        "e2e_pre_pub_publisher",
+      ],
+    },
+    {
+      orgShort: "dev-bas-p-hub",
+      url: "https://dev-bas-p-hub.mapsdevext.arcgis.com",
+      users: [
+        "e2e_bas_priv_admin",
+        "e2e_bas_priv_collaborator",
+        "e2e_bas_priv_publisher",
+      ],
+    },
+    {
+      orgShort: "dev-bas-hub",
+      url: "https://dev-bas-hub.mapsdevext.arcgis.com",
+      users: [
+        // "e2e_bas_pub_admin",
+        "e2e_bas_pub_collaborator",
+        "e2e_bas_pub_publisher",
+      ],
+    },
+  ],
+  admins: [
+    { orgShort: "dev-bas-hub", user: "dev_bas_hub_admin", env: "devext" },
+    {
+      orgShort: "dev-bas-a-hub",
+      user: "dev_bas_a_hub_admin",
+      env: "devext",
+    },
+    {
+      orgShort: "dev-bas-p-hub",
+      user: "dev_bas_p_hub_admin",
+      env: "devext",
+    },
+    { orgShort: "dev-pre-hub", user: "dev_pre_hub_admin", env: "devext" },
+    {
+      orgShort: "dev-pre-hub-c",
+      user: "dev_pre_hub_c_admin",
+      env: "devext",
+    },
+    {
+      orgShort: "dev-pre-a-hub",
+      user: "dev_pre_a_hub_admin",
+      env: "devext",
+    },
+    {
+      orgShort: "dev-pre-a-hub-c",
+      user: "dev_pre_a_hub_c_admin",
+      env: "devext",
+    },
+    {
+      orgShort: "dev-pre-p-hub",
+      user: "dev_pre_p_hub_admin",
+      env: "devext",
+    },
+    {
+      orgShort: "dev-pre-p-hub-c",
+      user: "dev_pre_p_hub_c_admin",
+      env: "devext",
+    },
+    { orgShort: "prod-bas-hub", user: "prod_bas_hub_admin", env: "prod" },
+    {
+      orgShort: "prod-bas-a-hub",
+      user: "prod_bas_a_hub_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-bas-p-hub",
+      user: "prod_bas_p_hub_admin",
+      env: "prod",
+    },
+    { orgShort: "prod-pre-hub", user: "prod_pre_hub_admin", env: "prod" },
+    {
+      orgShort: "prod-pre-hub-c",
+      user: "prod_pre_hub_c_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-pre-a-hub",
+      user: "prod_pre_a_hub_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-pre-a-hub-c",
+      user: "prod_pre_a_hub_c_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-pre-p-hub",
+      user: "prod_pre_p_hub_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-pre-p-hub-c",
+      user: "prod_pre_p_hub_c_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-test-act",
+      user: "prod_test_act_admin",
+      env: "prod",
+    },
+    {
+      orgShort: "prod-test-act-c",
+      user: "prod_test_act_community_admin",
+      env: "prod",
+    },
+    { orgShort: "qa-bas-hub", user: "qa_bas_hub_admin", env: "qaext" },
+    {
+      orgShort: "qa-bas-m-hub",
+      user: "qa_bas_m_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-bas-ma-hub",
+      user: "qa_bas_ma_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-bas-a-hub",
+      user: "qa_bas_a_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-bas-p-hub",
+      user: "qa_bas_p_hub_admin",
+      env: "qaext",
+    },
+    { orgShort: "qa-pre-hub", user: "qa_pre_hub_admin", env: "qaext" },
+    {
+      orgShort: "qa-pre-hub-c",
+      user: "qa_pre_hub_c_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-m-hub",
+      user: "qa_pre_m_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-m-hub-c",
+      user: "qa_pre_m_hub_c_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-a-hub",
+      user: "qa_pre_a_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-a-hub-c",
+      user: "qa_pre_a_hub_c_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-p-hub",
+      user: "qa_pre_p_hub_admin",
+      env: "qaext",
+    },
+    {
+      orgShort: "qa-pre-p-hub-c",
+      user: "qa_pre_p_hub_c_admin",
+      env: "qaext",
+    },
+  ],
+};